04-multimodal/pyproject.toml
[project]
name = "financial-advisor-bot"
version = "0.1.0"
description = "LLM-based Telegram bot for personal finance tracking"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "vosk>=0.3.45",
    "pydub>=0.25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

